small clean up of the root package. adding test_desk to test subpackage in integration with the exchange.
